Enabled per-thread diag context properties. Allow using system TID in logs. Fixed to-memory logging.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@32389 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/corelib/ncbidiag.inl
@@ -121,8 +121,8 @@ private:
     typedef list<string> TPrefixList;
     TPrefixList m_PrefixList;
 
-    typedef int TTID;
-    typedef int TPID;
+    typedef Uint8 TTID;
+    typedef Uint8 TPID;
     // Cached process ID
     TPID m_PID;
     // Cached thread ID
